requirements.txt
asgiref==3.2.7
backcall==0.1.0
beautifulsoup4==4.9.0
decorator==4.4.2
Django==3.0.5
django-bootstrap-modal-forms==1.5.0
django-bootstrap4==1.1.1
django-environ==0.4.5
entrypoints==0.3
flake8==3.7.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
mccabe==0.6.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
pytz==2019.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.1.9
